feat: Fix a bug where the lyrics and buttons were not changing
In SolidJS, you cannot destructure props. This was causing the props to be passed
in a non-reactive way. Hence the lack of updates.

Here's to 1 hour wasted because I didn't read the docs.

Fixes #22

See more here...
("Unlike in some other frameworks, you cannot use object destructuring on the props of a component. This is because the props object, behind the scenes, relies on Object getters to lazily retrieve values. Using object destructuring breaks the reactivity of props."

https://www.solidjs.com/guides/rendering#:~:text=Unlike%20in%20some%20other%20frameworks%2C%20you%20cannot%20use%20object%20destructuring%20on%20the%20props%20of%20a%20component.%20This%20is%20because%20the%20props%20object%2C%20behind%20the%20scenes%2C%20relies%20on%20Object%20getters%20to%20lazily%20retrieve%20values.%20Using%20object%20destructuring%20breaks%20the%20reactivity%20of%20props.)