feat: Implement explicit 'NowPlaying' functionality.
Previously, the now playing track would be the first item in the queue. However,
users need more fine-grained control over when to start playing or not. This feature
implemented makes it so that a user has to explicitly set a track to 'now playing' by either,
clicking on 'Play Now' or 'Play Next'

NB: Also, avoid the 'Bang operator (!) when working with possibly undefined data. It cost me a day
of looking for a breaking change.