enhancement: Implement wireless casting support (#119)
* chore: Rename project page to present

* enhancement: Bootstrap the PresentationAPI composable

* enhancement: Futher integrate the PresentationAPI

Currently, the "Present" page is being opened. However, the connection is disconneting immediately after creation.
This is preventing any data flow between the two pages.

* enhancement: ProjectionPayload can be sent to the presentation screen

Hurray! One step is done. Data can be sent from the controller to the receiver.
A few things remain:
* When the page is refreshed, the connection is lost. There should be a way to resume the connection.
* The connection state is not reflected on the application; the projection button and display buttons are not showing the new state.

* refactor: Further abstract usePresentationAPI

* refactor: Convert event listeners to bindings on usePresentationAPI

* enhancement: Implement SolidJS signal to capture connection state.

Since the object returned by the PresentationAPI is not reactive, I opted to manually
set its state.

I hope that this will not blow up in my face when a projection session is launched but fails to connect :)

* refactor: Ignore TSLint/ESLint errors on usePresentationAPI composable