fix(presentation): Fix the start and reconnect presentation methods to throw an error
When the startPresentation and reconnectPresentation methods fail, they were only logging an error. This
would cause the calling function to continue execution and eventually, set the isConnected state to 'true'
even though it failed. This fix addresses that by throwing the relevant errors, thus preventing illegal
setting of the isConnected signal.

Closes #120