enhancement: Implement wireless casting via the WindowManagementAPI
While the PresentationAPI is a superior API (in terms of the functionality required)
for this feature, it has one major drawback; that is that it does not support multiple
casting instances being controlled by from single tab. Until that feature is implemented,
we shall move back to the WindowManagementAPI. The composable has been refactored to ensure that it
is easy to replace with the PresentationAPI should need be.

Closes #122

**Changelog**
* Changed the `/present` route to have an optional unique ID to differentiate presentation sessions.
* Added preloading images to the UI.
* Added a robots.txt file.
* Add a banner to display when the application is offline.
* Update tests to reflect changes in UI and functionality.